feat: v0.6.0 - Paper 1 hardening complete: Qwen2.5-1.5B modern replication, systems implications, reproducibility fixes
SUMMARY
- Modern model replication: Qwen/Qwen2.5-1.5B (1.5B, Sep 2024, RoPE+GQA+SwiGLU)
  validated all hypotheses: H1 (+0.88 nats), H2 (-0.62 nats), H3 (-0.07 nats).
  4 models, 3 architecture families (GPT-2, GPT-Neo, Qwen), 82M-1.5B scale.
- Systems implications section (new §13): recursive planning drift, hallucinated
  refinement, synthetic contamination, spec degradation. Carefully scoped.
- BFloat16 fix: .float() before .numpy() for modern bf16 model compatibility.
- Reproducibility fixes: numpy==1.26.4 pin, verify_manifest.py, run_all_experiments.sh,
  Dockerfile CMD fixed, generate_insights_pdf.py (10-page report).
- Manuscript: 4-model Table 3, 4-panel figures, updated abstract/discussion/conclusion.
- Paper 1 unlock: 10/11 conditions met. Only arXiv submission remains.

+++ experiments/real_lm_experiment.py
@@ -394,7 +394,7 @@ def run_real_lm_experiment() -> list[dict]:
             return -100.0
         t = torch.tensor([ids[-256:]]).to(device)  # use last 256 tokens as context
         out = model(t)
-        lp = torch.nn.functional.log_softmax(out.logits[:, :-1, :], dim=-1)
+        lp = torch.nn.functional.log_softmax(out.logits[:, :-1, :].float(), dim=-1)
         per_tok = lp.gather(2, t[:, 1:].unsqueeze(-1)).squeeze(-1)
         return float(per_tok.mean())
 
@@ -423,7 +423,7 @@ def run_real_lm_experiment() -> list[dict]:
         """
         ctx = torch.tensor([context_ids[-64:]]).to(device)
         out = model(ctx)
-        lp = torch.nn.functional.log_softmax(out.logits[0, -1, :], dim=-1).cpu().numpy()
+        lp = torch.nn.functional.log_softmax(out.logits[0, -1, :].float(), dim=-1).cpu().numpy()
 
         rng = np.random.default_rng(99)
         vocab_size = tokenizer.vocab_size
